вынесена бизнес логика в presenter
